separate benchmarks from unit tests.
# Conflicts:
#	protocol_driver_test.cc
